# services/menu/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Menu Service API",
        "description": "Restaurant menu management service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Menu Service"
        }
    ],
    "paths": {
...
        "/v1/menu/validation": {
            "post": {
                "summary": "Validate menu items",
                "description": "Validates a list of menu items for availability and pricing",
                "operationId": "validateMenuItems",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "itemId",
                                                "qty",
                                                "unitPrice"
                                            ],
                                            "properties": {
                                                "itemId": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "unitPrice": {
                                                    "type": "string",
                                                    "example": "12.99"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "totalPrice": {
                                            "type": "string"
                                        },
                                        "traceId": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}